Amend TokenService.cs + few fixes
Amends TokenService.cs to correctly assign GivenName
and FamilyName claims + sets expiration to 5 minutes.
Creates a mat-card when logged in which displays
email and access token.
Now, on app.component.ts ngOnInit when user is loaded,
it will no longer display an error message in the
console and instead removes any JWT still in localStorage.